### 2020-10-15 #67 @whedon
[PRE REVIEW]: Causal.jl: A Modeling and Simulation Framework for Causal Models
**Submitting author:** @zekeriyasari (<a href="http://orcid.org/0000-0003-4070-9666">Zekeriya SARI</a>)
**Repository:** <a href="https://github.com/zekeriyasari/Causal.jl" target ="_blank">https://github.com/zekeriyasari/Causal.jl</a>
**Version:**
