### 2020-10-17 #67 @zekeriyasari
@matbesancon I renamed the subfolder "proceedings" to "paper".
