### 2020-10-19 #67 @matbesancon
It seems like you modified the template a lot. Could you keep the paper structure as-is? The compilation phase relies on it.
Also, https://github.com/zekeriyasari/Causal.jl/blob/juliacon/paper/paper.yml#L21
it seems like you are using a different file, Ref.bib, and not refs.bib
