+++ Project.toml
@@ -29,7 +29,7 @@ DifferentialEquations = "6.3"
 DocStringExtensions = "0.8"
 FileIO = "1.0"
 GraphPlot = "0.3, 0.4"
-Interpolations = "0.12"
+Interpolations = "0.12, 0.13"
 JLD2 = "0.1, 0.2"
 LightGraphs = "1.3"
 NLsolve = "4.3"
