### 2020-10-21 d35dca1
Deleted proceedings folder.
