+++ proceedings/paper.bib
@@ -1,154 +1,0 @@
-@article{elmqvist1978structured,
-  title={A structured model language for large continuous systems},
-  author={Elmqvist, Hilding},
-  journal={PhD Thesis TFRT-1015},
-  year={1978},
-  publisher={Department of Automatic Control, Lund Institute of Technology (LTH)}
-}
-@inproceedings{nytsch2006advanced,
-  title={Advanced modeling and simulation techniques in MOSILAB: A system development case study},
-  author={Nytsch-Geusen, Christoph and Ernst, Thilo and Nordwig, Andr{\'e} and Schwarz, Peter and Schneider, Peter and Vetter, Matthias and Wittwer, Christof and Holm, Andreas and Nouidui, Thierry and Leopold, J{\"u}rgen and others},
-  booktitle={Proceedings of the 5th International Modelica Conference, Arsenal Research, Wien},
-  year={2006}
-}
-@inproceedings{zimmer2008introducing,
-  title={Introducing Sol: A general methodology for equation-based modeling of variable-structure systems},
-  author={Zimmer, Dirk},
-  booktitle={Proceedings of the 6th International Modelica Conference},
-  pages={47--56},
-  year={2008},
-  organization={Citeseer}
-}
-@article{mosterman2002hybrsim,
-  title={HYBRSIM—a modelling and simulation environment for hybrid bond graphs},
-  author={Mosterman, Pieter J},
-  journal={Proceedings of the Institution of Mechanical Engineers, Part I: Journal of Systems and Control Engineering},
-  volume={216},
-  number={1},
-  pages={35--46},
-  year={2002},
-  publisher={SAGE Publications Sage UK: London, England}
-}
-@inproceedings{van2001variables,
-  title={Variables and equations in hybrid systems with structural changes},
-  author={Van Beek, DA},
-  booktitle={Proc. 13th European Simulation Symposium, Marseille},
-  pages={30--34},
-  year={2001},
-  organization={Citeseer}
-}
-@inproceedings{giorgidze2009higher,
-  title={Higher-order non-causal modelling and simulation of structurally dynamic systems},
-  author={Giorgidze, George and Nilsson, Henrik},
-  booktitle={Proceedings of the 7th International Modelica Conference; Como; Italy; 20-22 September 2009},
-  number={043},
-  pages={208--218},
-  year={2009},
-  organization={Link{\"o}ping University Electronic Press}
-}
-@article{pfeiffer2012pysimulator,
-  title={PySimulator--A simulation and analysis environment in Python with plugin infrastructure},
-  author={Pfeiffer, Andreas and Hellerer, Matthias and Hartweg, Stefan and Otter, Martin and Reiner, Matthias},
-  year={2012}
-}
-@article{rackauckas2017differentialequations,
-  title={Differentialequations. jl--a performant and feature-rich ecosystem for solving differential equations in julia},
-  author={Rackauckas, Christopher and Nie, Qing},
-  journal={Journal of Open Research Software},
-  volume={5},
-  number={1},
-  year={2017},
-  publisher={Ubiquity Press}
-}
-@article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM}
-}
-@book{matei2012modeling,
-  title={Modeling methodologies and simulation for dynamical systems},
-  author={Matei, Ion and Bock, Conrad and Matei, Ion},
[…]
