+++ proceedings/tex/README.md
@@ -1,1 +1,0 @@
-This repo is for the JuliaCon2020 Proceedings.

[…]
