### 2020-10-21 d53c77f
Added revised paper folder.
