+++ paper/paper.tex
@@ -1,0 +1,430 @@
+
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\setcounter{page}{1}
+
+% Packages added to the template
+\usepackage{subfig}
+\usepackage{amsmath}
+\usepackage{tikz}
+\usepackage{cite}
+
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+    This paper introduces a modeling and simulation framework, Causal.jl,  that enables fast and effective system simulations and online and offline data analyzes. Causal.jl adopts a causal modeling approach in which a model consists of components that process data and the connections that transfer the data flowing between these components. The framework developed makes it possible to simulate discrete time or continuous time, static or dynamical systems. In particular, it is possible to simulate dynamical systems modeled by various types of equations such as the ordinary, random ordinary, stochastic, delayed differential, differential-algebraic equations, and discrete-time difference equations. During the simulation, the data flowing through the connections can be processed online and offline, and specialized analyzes can be performed. These analyzes can also be enriched with plugins that can be easily defined using the standard Julia library or various Julia packages. The simulation is performed by evolving the model components between sampling time intervals individually and in parallel. The independent evolution of the components allows the simulation of the models consisting of the components represented by different mathematical equations, while the parallel evolution of components increases the simulation performance.
+
+\end{abstract}
+
+\section{Introduction}
+
+Numerical simulations can be expressed as solving the mathematical equations derived from modeling physical systems. Based on the system's properties at hand and abstraction level in the modeling, mathematical equations may be ordinary, stochastic, delay differential, or difference equations. High-speed performance and the ability to offer useful analysis tools are other typical features expected from an effective simulation environment.
+
+Many simulation environments are available for numerical analysis of systems\cite{elmqvist1978structured,nytsch2006advanced,zimmer2008introducing,mosterman2002hybrsim,van2001variables,giorgidze2009higher,pfeiffer2012pysimulator,simulink}. They are capable of allowing simulations that are represented by ordinary differential equations and differential algebraic equations, mostly. This is restrictive given the variety of mathematical equations that can be derived from the modeling\cite{rackauckas2017differentialequations}. Besides, many of the existing simulation environments lack modern computational methods such as parallel computing.
+
[…]
