+++ paper/figures/NetworkModel-WattzStrogatz/network_wattz_strogratz.tex
@@ -1,0 +1,106 @@
+\documentclass[margin=5mm]{standalone}
+\usepackage[utf8]{inputenc}
+\usepackage{tikz}
+
+\usetikzlibrary{calc, positioning}
+\usetikzlibrary{arrows.meta}
+\usetikzlibrary{matrix}
+\usetikzlibrary{shadows}
+\usepgflibrary{shapes.misc}
+\usepgflibrary{{shapes.geometric}}
+
+\pgfdeclarelayer{shadow} 
+\pgfsetlayers{shadow,main}
+\def\shadowradius{3pt}
+
+
+\def\mw{1.25cm}
+\def\mh{1cm}
+\def\trianglecoordinate{2mm}
+
+\newcommand\drawshadowbis[1]{
+    \begin{pgfonlayer}{shadow}
+        \fill[inner color=black,outer color=white] ($(#1.south west)$) circle (\shadowradius);
+        \fill[inner color=black ,outer color=white] ($(#1.north west)$) circle (\shadowradius);
+        \fill[inner color=black ,outer color=white] ($(#1.south east)$) circle (\shadowradius);
+        \fill[inner color=black,outer color=white] ($(#1.north east)$) circle (\shadowradius);
+        \fill[ top color=black, bottom color=white] ($(#1.south west)+((0,-\shadowradius)$) rectangle ($(#1.south east)$);
+        \fill[left color=black,right color=white] ($(#1.south east)$) rectangle ($(#1.north east)+((\shadowradius,0)$);
+        \fill[bottom color=black,top color=white] ($(#1.north west)$) rectangle ($(#1.north east)+((0,\shadowradius)$);
+        \fill[right color=black,left color=white] ($(#1.south west)$) rectangle ($(#1.north west)+(-\shadowradius,0)$);
+    \end{pgfonlayer}
+    }
+
+\tikzstyle{component} = [draw, fill=white, minimum width=\mw, minimum height=\mh, align=center]
+
+\tikzset{
+    border/.style = { 
+        draw, rectangle, minimum width=\mw, minimum height=\mh, ultra thick, align=center
+    },
+    Component/.pic = {
+        \node [border, font=\large](-edge){#1}; 
+        \draw[thick] ([xshift=\trianglecoordinate] -edge.south) -- ([yshift=\trianglecoordinate] -edge.south);
+        \draw[thick] ([xshift=-\trianglecoordinate] -edge.south) -- ([yshift=\trianglecoordinate] -edge.south);
+        \draw[thick] (-edge.south) |- ++(-3mm, -4mm) node[xshift=-2mm, yshift=-1mm] {T}; 
+    },
+}
+
+\tikzset{
+    clockborder/.style = { 
+        trapezium, trapezium angle=60, minimum width=1cm, draw, ultra thick
+    },
+    Clock/.pic = {
+        \node [clockborder, shape border rotate=-180](-clockedge){#1};
+        \draw[very thick] (-clockedge.east) -- ++(2cm, 0cm);
+        \def\sft{0.5}
+        \foreach \x in {0, 0.5, 1, 1.5}{
+            \draw[very thick] (\x + \sft, 0.1) -| ++(0.25cm, 0.25cm) -| ++ (0.25cm, -0.25cm);
+        }
+    },
+}
+
+\begin{document}
+    \begin{tikzpicture}[every node/.style={font=\Large}]
+        % Place the blocks 
[…]
