+++ paper/figures/CoupledChuaSimulation/odesystem.tex
@@ -1,0 +1,11 @@
+\documentclass[convert={density=900,size=1080x800,outext=.png}]{standalone}
+\usepackage[utf8]{inputenc}
+\usepackage{tikz}
+
+\begin{document}
+    \begin{tikzpicture}[every node/.style={font=\Huge}]
+        \node [](p){\includegraphics{odesystem_raw.png}};
+        \node [xshift=0.75cm, yshift=-5.5cm]{$x_1$};
+        \node [xshift=-8.5cm, yshift=0.25cm]{$y_1$};
+    \end{tikzpicture}
+\end{document}

[…]
