### 2020-10-21 #67 @zekeriyasari
@matbesancon I revised the paper files. It has been 13 minutes and no error has been thrown yet. I think the errors are fixed now.
