### 2021-01-27 #67 @arfon
@matbesancon - I think the fact that there is a `paper.md` and `paper.tex` is confusing Whedon. Can you delete the `paper.md` file and try again?
