### 2021-01-28 #67 @zekeriyasari
@matbesancon I removed `paper.md`. Seems like the build is OK.
