### 2021-02-11 #71 @YingboMa
Links for JuliaCon conflict of interest policy and JuliaCon code of conduct don't seem to work.
