### 2021-02-11 #71 @matbesancon
Sorry yes we are fixing some of the links now

https://juliacon.github.io/proceedings-guide/reviewer/#conflict_of_interest
https://juliacon.github.io/proceedings-guide/reviewer/#code_of_conduct
