+++ Project.toml
@@ -30,7 +30,7 @@ DocStringExtensions = "0.8"
 FileIO = "1.0"
 GraphPlot = "0.3, 0.4"
 Interpolations = "0.12, 0.13"
-JLD2 = "0.1, 0.2, 0.3"
+JLD2 = "0.1, 0.2, 0.3, 0.4"
 LightGraphs = "1.3"
 NLsolve = "4.3"
 Plots = "0.24, 0.29, 1.2"
