### 2021-03-11 #71 @cscherrer
Hi, I've annotated the PDF with some feedback here:
[causal-juliacon-annotated.pdf](https://github.com/JuliaCon/proceedings-review/files/6125944/causal-juliacon-annotated.pdf)

Some of this is outside my area of expertise, but I hope the feedback can be helpful in strengthening the paper. Please let me know if any of my handwriting is too messy to read, I can type it out if that's the case :)
