### 2021-03-12 #71 @zekeriyasari
Thank you for your feedback @cscherrer . The feedback will definitely be helpful to improve the paper. I did not have any difficulty with reading your handwriting. All are clear.
