+++ Project.toml
@@ -36,4 +36,4 @@ NLsolve = "4.3"
 Plots = "0.24, 0.29, 1.2"
 ProgressMeter = "1.1"
 Sundials = "3.3, 4.2"
-julia = "1.3"
+julia = "1.4"
