+++ Project.toml
@@ -24,16 +24,4 @@ Sundials = "c3572dad-4567-51f8-b174-8c6c989267f4"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [compat]
-DataStructures = "0.15, 0.17, 0.18"
-DifferentialEquations = "6.3"
-DocStringExtensions = "0.8"
-FileIO = "1.0"
-GraphPlot = "0.3, 0.4"
-Interpolations = "0.12, 0.13"
-JLD2 = "0.1, 0.2, 0.3, 0.4"
-LightGraphs = "1.3"
-NLsolve = "4.3"
-Plots = "0.24, 0.29, 1.2"
-ProgressMeter = "1.1"
-Sundials = "3.3, 4.2"
 julia = "1.4"
