### 2021-03-12 fcb66bb
CompatHelper: add new compat entry for "Sundials" at version "4.4"
