### 2021-03-12 eb199e5
CompatHelper: add new compat entry for "GraphPlot" at version "0.4"
