### 2021-03-12 5c6f39f
CompatHelper: add new compat entry for "LightGraphs" at version "1.3"
