+++ Project.toml
@@ -24,4 +24,5 @@ Sundials = "c3572dad-4567-51f8-b174-8c6c989267f4"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [compat]
+LightGraphs = "1.3"
 julia = "1.4"
