### 2021-03-12 a26bf85
CompatHelper: add new compat entry for "NLsolve" at version "4.5"
