### 2021-03-12 da64732
CompatHelper: add new compat entry for "DifferentialEquations" at version "6.16"
