### 2021-03-12 071e2cb
CompatHelper: add new compat entry for "DataStructures" at version "0.18"
