### 2021-03-12 94c8720
CompatHelper: add new compat entry for "ProgressMeter" at version "1.5"
