### 2021-03-12 a1b8709
CompatHelper: add new compat entry for "DocStringExtensions" at version "0.8"
