### 2021-03-12 bad5bb2
CompatHelper: add new compat entry for "Plots" at version "1.10"
