### 2021-03-12 9c062e9
CompatHelper: add new compat entry for "FileIO" at version "1.6"
