### 2021-03-12 0097de3
CompatHelper: add new compat entry for "JLD2" at version "0.4"
