### 2021-03-12 96f17a5
CompatHelper: add new compat entry for "Interpolations" at version "0.13"
