+++ Project.toml
@@ -24,4 +24,16 @@ Sundials = "c3572dad-4567-51f8-b174-8c6c989267f4"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [compat]
+DataStructures = "0.18"
+DifferentialEquations = "6.16"
+DocStringExtensions = "0.8"
+FileIO = "1.6"
+GraphPlot = "0.4"
+Interpolations = "0.13"
+JLD2 = "0.4"
+LightGraphs = "1.3"
+NLsolve = "4.5"
+Plots = "1.10"
+ProgressMeter = "1.5"
+Sundials = "4.4"
 julia = "1.4"
