+++ paper/ref.bib
@@ -156,3 +156,10 @@
   url          = {https://doi.org/10.5281/zenodo.889971},
   howpublished = {"\url{https://github.com/JuliaGraphs/LightGraphs.jl}"}
 }
+@misc{dagger,
+  title        = {Dagger.jl - A framework for out-of-core and parallel execution},
+  url          = {https://github.com/JuliaParallel/Dagger.jl},
+  howpublished = {"\url{https://github.com/JuliaParallel/Dagger.jl}"},
+  urldate      = {2021-04-25},
+  year         = {2021}
+}
