### 2021-04-29 cdac1e1
Added methods to re-simulate a model.
