+++ src/Causal.jl
@@ -223,8 +223,10 @@ export @defmodel,
     Branch, 
     addnode!, 
     getnode, 
+    getcomponent,
     addbranch!,
     getbranch, 
+    getlinks,
     deletebranch!, 
     signalflow,
     troubleshoot
