+++ src/models/model.jl
@@ -115,6 +115,13 @@ Node(component:Gain(gain:1.0, input:Inport(numpins:1, eltype:Inpin{Float64}), ou
 getnode(model::Model, idx::Int) = model.nodes[idx]
 getnode(model::Model, label) = filter(node -> node.label === label, model.nodes)[1]
 
+"""
+    $(SIGNATURES)
+
+Returns the component of `model` corresponding to `specifier` that can be either index or label of the component.
+"""
+getcomponent(model::Model, specifier) = getnode(model, specifier).component
+
 function register(taskmanager, component)
     triggerport, handshakeport = taskmanager.triggerport, taskmanager.handshakeport
     triggerpin, handshakepin = Outpin(), Inpin{Bool}()
@@ -144,6 +151,13 @@ end
 getbranch(model::Model, nodepair::Pair{Int, Int}) = filter(branch -> branch.nodepair == nodepair, model.branches)[1]
 getbranch(model::Model, nodepair::Pair{Symbol, Symbol}) = 
     getbranch(model, getnode(model, nodepair.first).idx => getnode(model, nodepair.second).idx)
+
+""" 
+    $(SIGNATURES)
+
+Returns the links of `model` corresponding to the `pair` which can be a pair of integers or symbols to specify the source and destination nodes of the branch.
+"""
+getlinks(model::Model, pair) = getbranch(model, nodepair).links
 
 """
     deletebranch!(model::Model, branch::Branch)
@@ -413,43 +427,28 @@ function checkchannels(model)
 end
 
 ##### Model initialization
+
 """
-    initialize!(model::Model)
+    $(SIGNATURES)
 
 Initializes `model` by launching component task for each of the component of `model`. The pairs component and component tasks are recordedin the task manager of the `model`. The `model` clock is [`set!`](@ref) and the files of [`Writer`](@ref) are openned.
 """
 function initialize!(model::Model)
-    taskmanager = model.taskmanager
-    pairs = taskmanager.pairs
-    nodes = model.nodes
-
     # NOTE: Tasks to make the components be triggerable are launched here.
     # The important point here is that the simulation should be cancelled if an error is thrown in any of the tasks 
     # launched here. This is done by binding the task to the chnnel of the trigger link of the component. Hrence the 
     # lifetime of the channel of the link connecting the component to the taskmanger is determined by the lifetime of 
     # the task launched for the component. To cancel the simulation and report the stacktrace the task is `fetch`ed. 
-    for node in nodes 
-        component = node.component
-        link = whichlink(taskmanager, component)  # Link connecting the component to taskmanager. 
-        task = launch(component)    # Task launched to make `componnent` be triggerable.
-        bind(link.channel, task)    # Bind the task to the channel of the link. 
-        pairs[component] = task 
-    end
+    bind!(model)
 
     # Turn on clock model clock if it is running. 
-    if isoutoftime(model.clock)
-        msg = "Model clock is out of time. Its current time $(model.clock.t) should be less than its final time "
-        msg *= "$(model.clock.tf). Resettting the model clock to its defaults."
-        @warn msg
-        set!(model.clock)
-    end
-    isrunning(model.clock) || set!(model.clock)  
+    initclock!(model)
+
+    # Clean the model 
+    clean!(model)
     
     # Open the files, GUI's for sink components. 
-    foreach(node -> open(node.component), filter(node->isa(node.component, AbstractSink), model.nodes))
-
-    # Return the model back.
-    model
+    opensinks!(model)
 end
 
 # Find the link connecting `component` to `taskmanager`.
@@ -460,6 +459,76 @@ function whichlink(taskmanager, component)
[…]
