### 2021-04-29 2b8de36
Revised dynamical system interpolant.
