+++ src/components/systems/dynamicalsystems/discretesystems.jl
@@ -42,7 +42,7 @@ Here, `MyDiscreteSystem` has `N` parameters. `MyDiscreteSystem` is represented b
     New discrete system must be a subtype of `AbstractDiscreteSystem` to function properly.
 
 # Example 
-```jldoctest 
+```julia 
 julia> @def_discrete_system mutable struct MyDiscreteSystem{RH, RO, IP, OP} <: AbstractDiscreteSystem 
        α::Float64 = 1. 
        β::Float64 = 2. 
@@ -61,41 +61,17 @@ julia> ds.input
 ```
 """
 macro def_discrete_system(ex) 
-    #= NOTE: 
-        Generate the parameter names TR, HS, CB, ID, MA, MK, SA, SK, AL, IT  of types of trigger, handshake, callbacks, 
-        id, modelargs, modelkwargs, solverargs, solverkwargs, alg, integrator by using `gensym` to avoid duplicate type 
-        parameter names so that the users can parametrizde their types as 
-        
-        @def_discrete_system mutable struct MyDiscreteSystem{RH, RO, ST, IP, TR} <: AbstractDiscreteSystem
-            righthandside::RH
-            readout::RO 
-            state::ST 
-            input::IP 
-            output::TR
-        end
-                
-        Note that the parameter name of output is TR. But, since the parameter name TR of trigger of the compnent 
-        is generated by `gensym`, we get rid of duplicate parameter names.
-    =#
     checksyntax(ex, :AbstractDiscreteSystem)
-    # TR, HS, CB, ID, MA, MK, SA, SK, AL, IT = [gensym() for i in 1 : 10]
-    fields = quote
-        trigger::$(TRIGGER_TYPE_SYMBOL) = Inpin()
-        handshake::$(HANDSHAKE_TYPE_SYMBOL) = Outpin{Bool}()
-        callbacks::$(CALLBACKS_TYPE_SYMBOL) = nothing
-        name::Symbol = Symbol()
-        id::$(ID_TYPE_SYMBOL)= Causal.uuid4()
-        t::Float64 = 0.
-        modelargs::$(MODEL_ARGS_TYPE_SYMBOL) = () 
-        modelkwargs::$(MODEL_KWARGS_TYPE_SYMBOL) = NamedTuple() 
-        solverargs::$(SOLVER_ARGS_TYPE_SYMBOL) = () 
-        solverkwargs::$(SOLVER_KWARGS_TYPE_SYMBOL) = NamedTuple() 
-        alg::$(ALG_TYPE_SYMBOL) = Causal.FunctionMap()
-        integrator::$(INTEGRATOR_TYPE_SYMBOL) = Causal.construct_integrator(Causal.DiscreteProblem, input, righthandside, state, t,
-            modelargs, solverargs; alg=alg, modelkwargs=modelkwargs, solverkwargs=solverkwargs, numtaps=3)
-    end, [TRIGGER_TYPE_SYMBOL, HANDSHAKE_TYPE_SYMBOL, CALLBACKS_TYPE_SYMBOL, ID_TYPE_SYMBOL, MODEL_ARGS_TYPE_SYMBOL, MODEL_KWARGS_TYPE_SYMBOL, SOLVER_ARGS_TYPE_SYMBOL, SOLVER_KWARGS_TYPE_SYMBOL, ALG_TYPE_SYMBOL, INTEGRATOR_TYPE_SYMBOL]
-    _append_common_fields!(ex, fields...)
-    deftype(ex)
+    appendcommonex!(ex)
+    foreach(nex -> appendex!(ex, nex), [
+        :( alg::$ALG_TYPE_SYMBOL = Causal.FunctionMap() ), 
[…]
