+++ src/components/systems/dynamicalsystems/ddesystems.jl
@@ -45,7 +45,7 @@ Here, `MyDDESystem` has `N` parameters. `MyDDESystem` is represented by the `rig
     New DDE system must be a subtype of `AbstractDDESystem` to function properly.
 
 # Example 
-```jldoctest 
+```julia 
 julia> _delay_feedback_system_cache = zeros(1)
 1-element Array{Float64,1}:
  0.0
@@ -83,29 +83,21 @@ julia> ds = MyDDESystem();
 """
 macro def_dde_system(ex)
     checksyntax(ex, :AbstractDDESystem)
-    fields = quote
-        trigger::$(TRIGGER_TYPE_SYMBOL) = Inpin()
-        handshake::$(HANDSHAKE_TYPE_SYMBOL) = Outpin{Bool}()
-        callbacks::$(CALLBACKS_TYPE_SYMBOL) = nothing
-        name::Symbol = Symbol()
-        id::$(ID_TYPE_SYMBOL) = Causal.uuid4()
-        t::Float64 = 0.
-        modelargs::$(MODEL_ARGS_TYPE_SYMBOL) = () 
-        modelkwargs::$(MODEL_KWARGS_TYPE_SYMBOL) = NamedTuple() 
-        solverargs::$(SOLVER_ARGS_TYPE_SYMBOL) = () 
-        solverkwargs::$(SOLVER_KWARGS_TYPE_SYMBOL) = NamedTuple() 
-        alg::$(ALG_TYPE_SYMBOL) = Causal.MethodOfSteps(Causal.Tsit5())
-        integrator::$(INTEGRATOR_TYPE_SYMBOL) = Causal.construct_integrator(
+    appendcommonex!(ex)
+    foreach(nex -> appendex!(ex, nex), [
+        :( alg::$ALG_TYPE_SYMBOL = Causal.MethodOfSteps(Causal.Tsit5()) ), 
+        :( integrator::$INTEGRATOR_TYPE_SYMBOL = Causal.construct_integrator(
             Causal.DDEProblem, input, (righthandside, history), state, t, modelargs, solverargs; 
             alg=alg, modelkwargs=(; 
             zip(
                 (keys(modelkwargs)..., :constant_lags, :dependent_lags), 
                 (values(modelkwargs)..., constlags, depslags))...
                 ), 
-            solverkwargs=solverkwargs, numtaps=3)
-    end, [TRIGGER_TYPE_SYMBOL, HANDSHAKE_TYPE_SYMBOL, CALLBACKS_TYPE_SYMBOL, ID_TYPE_SYMBOL, MODEL_ARGS_TYPE_SYMBOL, MODEL_KWARGS_TYPE_SYMBOL, SOLVER_ARGS_TYPE_SYMBOL, SOLVER_KWARGS_TYPE_SYMBOL, ALG_TYPE_SYMBOL, INTEGRATOR_TYPE_SYMBOL]
-    _append_common_fields!(ex, fields...)
-    deftype(ex)
+            solverkwargs=solverkwargs, numtaps=3) ) 
+        ])
+    quote 
+        Base.@kwdef $ex 
+    end |> esc 
 end
 
 ##### Define DDE system library.
