### 2021-05-01 d4b5d8c
Added Github CI workflow.
