### 2021-05-01 a83cf25
Added Github CI badge to README
