### 2021-05-01 4a69562
Removed travis and appveyor .yml files.
