+++ .travis.yml
@@ -1,43 +1,0 @@
-
-language: julia
-
-os:
-  - linux
-  # - osx
-
-julia:
-  - 1.5
-  - nightly
-
-addons:
-  apt:
-    packages:
-    - qt5-default
-
-env:
-  - GKSwstype="100"
-
-script: # the default script is equivalent to the following
-  - while sleep 30; do echo "still alive"; done &
-  - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
-  - julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
-
-notifications:
-  email: false
-
-after_success:
-  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder())'
-  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(process_folder())'
-
-jobs:
-  allow_failures:
-    - julia: nightly
-  fast_finish: true
-  include:
-    - stage: Documentation
-      julia: 1.5
-      script:
-        - while sleep 30; do echo "still alive"; done & 
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
-      after_success: skip
