+++ test/coverage.jl
@@ -3,7 +3,7 @@
 # NOTE: Before executing this script, run test suite of Causal as 
 # ] test --coverage Causal
 # 
-nothing
+
 using Coverage
 
 # process '*.cov' files
