### 2021-05-08 3a80268
Made Clock a generaic iterator. [skip ci]
