+++ src/models/model.jl
@@ -39,26 +39,23 @@ Constructs a `Model` with empty components. After the construction, components c
 !!! warning
     `Model`s are units that can be simulated. As the data flows through the branches i.e. input output busses of the components, its is important that the components must be connected to each other. See also: [`simulate!`](@ref)
 """
-struct Model{GR, ND, BR, CK, TM, CB}
+struct Model{GR, ND, BR, TM, CB}
     graph::GR
     nodes::ND
     branches::BR 
-    clock::CK
     taskmanager::TM
     callbacks::CB
     name::Symbol
     id::UUID
-    function Model(nodes::AbstractVector=[], branches::AbstractVector=[]; 
-        clock=Clock(0, 0.01, 1.), callbacks=nothing, name=Symbol())
+    function Model(nodes::AbstractVector=[], branches::AbstractVector=[]; callbacks=nothing, name=Symbol())
         graph = SimpleDiGraph()
         taskmanager = TaskManager()
-        new{typeof(graph), typeof(nodes), typeof(branches), typeof(clock), typeof(taskmanager),
-            typeof(callbacks)}(graph, nodes, branches, clock, taskmanager, callbacks, name, uuid4())
+        new{typeof(graph), typeof(nodes), typeof(branches), typeof(taskmanager),
+            typeof(callbacks)}(graph, nodes, branches, taskmanager, callbacks, name, uuid4())
     end
 end
 
-show(io::IO, model::Model) = print(io, "Model(numnodes:$(length(model.nodes)), ",
-    "numedges:$(length(model.branches)), timesettings=($(model.clock.t), $(model.clock.dt), $(model.clock.tf)))")
+show(io::IO, model::Model) = print(io, "Model(numnodes:$(length(model.nodes)), numedges:$(length(model.branches)))")
 
 
 ##### Addinng nodes and branches.
@@ -433,7 +430,7 @@ end
 
 Initializes `model` by launching component task for each of the component of `model`. The pairs component and component tasks are recordedin the task manager of the `model`. The `model` clock is [`set!`](@ref) and the files of [`Writer`](@ref) are openned.
 """
-function initialize!(model::Model)
+function initialize!(model::Model, clock::Clock)
     # NOTE: Tasks to make the components be triggerable are launched here.
     # The important point here is that the simulation should be cancelled if an error is thrown in any of the tasks 
     # launched here. This is done by binding the task to the chnnel of the trigger link of the component. Hrence the 
@@ -441,8 +438,8 @@ function initialize!(model::Model)
     # the task launched for the component. To cancel the simulation and report the stacktrace the task is `fetch`ed. 
     bind!(model)
 
-    # Turn on clock model clock if it is running. 
-    initclock!(model)
+    # Reset dynamical systems 
+    resetdynamicalsystems!(model, clock)
 
     # Clean the model 
     clean!(model)
@@ -482,17 +479,17 @@ function opensinks!(model::Model)
     model 
 end 
 
-# Initialze model clock 
-function initclock!(model::Model)
-    if isoutoftime(model.clock)
-        msg = "Model clock is out of time. Its current time $(model.clock.t) should be less than its final time "
-        msg *= "$(model.clock.tf). Resettting the model clock to its defaults."
-        @warn msg
-        set!(model.clock)
-    end
-    isrunning(model.clock) || set!(model.clock)  
-    model
-end
+# # Initialze model clock 
+# function initclock!(model::Model)
+#     if isoutoftime(model.clock)
+#         msg = "Model clock is out of time. Its current time $(model.clock.t) should be less than its final time "
[…]
