+++ src/components/sources/clock.jl
@@ -2,200 +2,29 @@
 
 import Base: iterate, take!, length
 
-Generator(t0, dt, tf) = 
-    Channel{promote_type(typeof(t0),typeof(dt),typeof(tf))}(channel -> foreach(t -> put!(channel, t), t0:dt:tf))
-
-"""
-    Clock(t::Real, dt::Real, tf::Real)
-
-Constructs a `Clock` with starting time `t`, final time `tf` and sampling inteval `dt`. When iterated, the `Clock` returns its current time. 
-
-!!! warning 
-    When constructed, `Clock` is not running. To take clock ticks from `Clock`, the `Clock` must be setted. See [`take!(clk::Clock)`](@ref) and [`set!`](@ref) 
-"""
 mutable struct Clock{T, CB}
-    t::T 
-    ti::T 
-    dt::T 
-    tf::T 
-    generator::Channel{T}
+    generator::T
     paused::Bool
     callbacks::CB
     name::Symbol
     uuid::UUID
-    Clock{T}(ti::T, dt::T, tf::T; callbacks::CB=nothing, name=Symbol()) where {T, CB} = 
-        new{T, CB}(ti, ti, dt, tf, Channel{T}(0), false, callbacks, name, uuid4())
+    function Clock(generator::T; callbacks::CB=nothing, name=Symbol()) where {T, CB} 
+        if hasmethod(iterate, Tuple{T})
+            new{T, CB}(generator, false, callbacks, name, uuid4())
+        else 
+            error("$gen is not iterable")
+        end 
+    end 
 end
-Clock(t::T, dt::T, tf::T; kwargs...) where T = Clock{T}(t, dt, tf; kwargs...)
-Clock(t=0., dt=0.01, tf=1.; kwargs...) = Clock(promote(t, dt, tf)...; kwargs...)
+Clock(ti::Real, dt::Real, tf::Real; kwargs...) = Clock(ti:dt:tf; kwargs...)
 
-show(io::IO, clk::Clock) = print(io, 
-    "Clock(t:$(clk.t), dt:$(clk.dt), tf:$(clk.tf), paused:$(clk.paused), isrunning:$(isrunning(clk)))")
+show(io::IO, clk::Clock) = print(io, "Clock(gen:$(clk.generator), paused:$(clk.paused))")
 
-##### Reading from clock
-"""
-    take!(clk::Clock)
-
-Takes a values from `clk`.
-
-# Example 
-```jldoctest
-julia> clk = Clock(0., 0.1, 0.5)
-Clock(t:0.0, dt:0.1, tf:0.5, paused:false, isrunning:false)
-
-julia> set!(clk)
-Clock(t:0.0, dt:0.1, tf:0.5, paused:false, isrunning:true)
-
-julia> for i in 0 : 5
-       @show take!(clk)
-       end
-take!(clk) = 0.0
-take!(clk) = 0.1
-take!(clk) = 0.2
-take!(clk) = 0.3
-take!(clk) = 0.4
-take!(clk) = 0.5
-```
-"""
-function take!(clk::Clock)
-    if ispaused(clk)
-        @warn "Clock is paused."
-        return clk.t
-    end
-    if isoutoftime(clk)
-        @warn "Clock is out of time."
-        return clk.t
-    end
-    ## 
-    ##  NOTE: If this code block is uncommented, a bug occurs
-    ##  The same clock times is sent multiple time.
-    ## 
-    # if !isrunning(clk)   
-    #     @warn "Clock is not running."
-    #     return clk.t
-    # end
-    clk.t = take!(clk.generator)
-    applycallbacks(clk)
-    clk.t
[…]
