### 2021-05-09 dbdc5c3
Nonuniform sampling intervals. Fixes #75
