+++ test/test_plugins/test_plugins.jl
@@ -9,11 +9,12 @@
     end 
 
     # Try equip a writer in a model.
-    model = Model(clock=Clock(0., 0.01, 10.)) 
+    model = Model() 
+    clock=Clock(0., 0.01, 10.)
     addnode!(model, SinewaveGenerator(), label=:gen)
     addnode!(model, Writer(buflen=50, plugin=MeanPlugin()), label=:writer)
     addbranch!(model, :gen => :writer)
-    simulate!(model)
+    simulate!(model, clock)
 
     # Test the simulation data
     data = read(getnode(model, :writer).component, flatten=false)
