### 2021-05-09 94644bd
Revised docstring.
