+++ src/plugins/essentials/lyapunov.jl
@@ -4,14 +4,25 @@
 using ChaosTools
 
 """
-    Lyapunov(;m::Int=15, J::Int=5, ni::Int=300, ts::Float64=0.01)
+    $TYPEDEF 
 
-Constructs a `Lyapunov` plugin. The [`process(plg::Lyapunov, x)`](@ref) function computes the maximum numerical Lyapunov exponents. `m` is the reconstruction dimension, `J` is the amount of delay in reconstruction, `ni` is the number of steps during transient steps and `ts` is the sampling time between samples of the input data `x`. See also: (https://juliadynamics.github.io/DynamicalSystems.jl/latest/chaos/nlts/)
+Constructs a `Lyapunov` plugin. The [`process(plg::Lyapunov, x)`](@ref) function computes the maximum numerical Lyapunov
+exponents. `m` is the reconstruction dimension, `J` is the amount of delay in reconstruction, `ni` is the number of steps
+during transient steps and `ts` is the sampling time between samples of the input data `x`. See also:
+(https://juliadynamics.github.io/DynamicalSystems.jl/latest/chaos/nlts/)
+
+# FieldS 
+
+    $TYPEDFIELDS
 """
 struct Lyapunov <: AbstractPlugin
+    "Dimension"
     m::Int
+    "Delay amount"
     J::Int
+    "Transient nmber"
     ni::Int
+    "Samping sampling"
     ts::Float64
 end
 Lyapunov(;m::Int=15, J::Int=5, ni::Int=300, ts::Float64=0.01) = Lyapunov(m, J, ni, ts)
@@ -20,7 +31,7 @@ show(io::IO, plg::Lyapunov) = print(io,
     "Lyapunov(embeddingdim:$(plg.m), numlags:$(plg.J), numiteration:$(plg.ni), samplingtime:$(plg.ts)")
 
 """
-    process(plg::Lyapunov, x)
+    $SIGNATURES
 
 Computes the maximum Lyapunov exponent of the input data `x`. 
 
@@ -49,47 +60,20 @@ function process(plg::Lyapunov, x)
     return val
 end
 
-# using NearestNeighbors
-# using LinearAlgebra
-# using Statistics
-# using LsqFit
-# import Base.log
-# process(plg::Lyapunov, x) = lyapunov(x, plg.m, plg.J, plg.ni, plg.ts)[2]
+# using NearestNeighbors using LinearAlgebra using Statistics using LsqFit import Base.log process(plg::Lyapunov, x) =
+# lyapunov(x, plg.m, plg.J, plg.ni, plg.ts)[2]
 
-# function reconstruct(x, m, J)
-#     N = length(x)
-#     M = N - (m - 1) * J
-#     X = zeros(m, M)
-#     for i = 1 : M
-#         data =  x[i : J : i + (m - 1)* J]
-#         X[:, i] = data
-#     end
-#     X
-# end
+# function reconstruct(x, m, J) N = length(x) M = N - (m - 1) * J X = zeros(m, M) for i = 1 : M data =  x[i : J : i + (m -
+#     1)* J] X[:, i] = data end X end
 
-# function knneighbours(X)
-#     tree = KDTree(X)
-#     [knn(tree, X[:, j], 2)[1][1] for j = 1 : size(X, 2)]
-# end
+# function knneighbours(X) tree = KDTree(X) [knn(tree, X[:, j], 2)[1][1] for j = 1 : size(X, 2)] end
 
 # distance(Xj, Xjhat) = norm(Xj - Xjhat)
 
 # Base.log(a::Vector) = log.(a)
 
-# function lyapunov(x, m, J, ni, ts)
[…]
