### 2021-05-09 de89a31
Revised docstrings of sources #78
