+++ src/components/sources/generators.jl
@@ -23,7 +23,7 @@ Here, `MySource` has `N` parameters, an `output` port and a `readout` function.
     `output` and `readout` are mandatory fields to define a new source. The rest of the fields are the parameters of the source.
 
 !!! warning 
-    `readout` must be a single-argument function, i.e. a fucntion of time `t`.
+    `readout` must be a single-argument function, i.e. a function of time `t`.
 
 !!! warning 
     New source must be a subtype of `AbstractSource` to function properly.
@@ -65,9 +65,13 @@ end
 
 ##### Define Sources library
 """
-    FunctionGenerator(; readout, output=Outport()) 
+    $TYPEDEF
 
 Constructs a generic function generator with `readout` function and `output` port.
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example 
 ```jldoctest 
@@ -80,96 +84,144 @@ julia> gen.readout(1.)
 ```
 """
 @def_source struct FunctionGenerator{RO, OP} <: AbstractSource 
+    "Readout function"
     readout::RO 
+    "Output port"
     output::OP = Outport(1)    
 end
 
- @doc raw"""
-    SinewaveGenerator(;amplitude=1., frequency=1., phase=0., delay=0., offset=0.)
+ """
+    $SIGNATURES
 
 Constructs a `SinewaveGenerator` with output of the form
 ```math 
-    x(t) = A sin(2 \pi f  (t - \tau) + \phi) + B
+    x(t) = A sin(2 \\pi f  (t - \\tau) + \\phi) + B
 ```
-where ``A`` is `amplitude`, ``f`` is `frequency`, ``\tau`` is `delay` and ``\phi`` is `phase` and ``B`` is `offset`.
+where ``A`` is `amplitude`, ``f`` is `frequency`, ``\\tau`` is `delay` and ``\\phi`` is `phase` and ``B`` is `offset`.
+
+# Fields 
+
+    $TYPEDFIELDS
 """
 @def_source struct SinewaveGenerator{RO, OP} <: AbstractSource
+   "Amplitude"
     amplitude::Float64 = 1.
+    "Frequency"
     frequency::Float64 = 1. 
+    "Phase"
     phase::Float64 = 0. 
+    "Delay in seconds"
     delay::Float64 = 0. 
+    "Offset"
     offset::Float64 = 0.
+    "Output port"
     output::OP = Outport()
+    "Readout function"
     readout::RO = (t, amplitude=amplitude, frequency=frequency, delay=delay, offset=offset) ->
         amplitude * sin(2 * pi * frequency * (t - delay) + phase) + offset 
 end
 
 
-@doc raw"""
-    DampedSinewaveGenerator(;amplitude=1., decay=-0.5, frequency=1., phase=0., delay=0., offset=0.)
+"""
+    $TYPEDEF
 
 Constructs a `DampedSinewaveGenerator` which generates outputs of the form 
 ```math 
-    x(t) = A e^{\alpha t} sin(2 \pi f (t - \tau) + \phi) + B
+    x(t) = A e^{\\alpha t} sin(2 \\pi f (t - \\tau) + \\phi) + B
 ```
-where ``A`` is `amplitude`, ``\alpha`` is `decay`, ``f`` is `frequency`, ``\phi`` is `phase`, ``\tau`` is `delay` and ``B`` is `offset`.
[…]
