### 2021-05-09 d984b06
Revised docstring for staticsystems #78
