+++ src/components/systems/staticsystems/staticsystems.jl
@@ -67,9 +67,13 @@ end
 ##### Define prototipical static systems.
 
 """
-    StaticSystem(; readout, input, output)
+    $TYPEDEF
 
 Consructs a generic static system with `readout` function, `input` port and `output` port.
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example 
 ```jldoctest 
@@ -80,20 +84,27 @@ julia> ss.readout(0., ones(2))
 ```
 """
 @def_static_system struct StaticSystem{RO, IP, OP} <: AbstractStaticSystem
+    "Readout function"
     readout::RO 
+    "Input port"
     input::IP 
+    "Output. May be an `Outport` of `Nothing`"
     output::OP
 end
 
 
-@doc raw"""
-    Adder(signs=(+,+))
+"""
+    $TYPEDEF
 
 Construts an `Adder` with input bus `input` and signs `signs`. `signs` is a tuplle of `+` and/or `-`. The output function `g` of `Adder` is of the form,
 ```math 
-    y = g(u, t) =  \sum_{j = 1}^n s_k u_k
+    y = g(u, t) =  \\sum_{j = 1}^n s_k u_k
 ```
 where `n` is the length of the `input`, ``s_k`` is the `k`th element of `signs`, ``u_k`` is the `k`th value of `input` and ``y`` is the value of `output`. The default value of `signs` is all `+`.
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example 
 ```jldoctest
@@ -104,21 +115,29 @@ true
 ```
 """
 @def_static_system struct Adder{S, IP, OP, RO} <: AbstractStaticSystem 
+    "Addition signs"
     signs::S = (+, +)
+    "Input port"
     input::IP = Inport(length(signs))
+    "Output port"
     output::OP = Outport()
+    "Readout function"
     readout::RO = (u, t, signs=signs) -> sum([sign(val) for (sign, val) in zip(signs, u)])
 end
 
 
-@doc raw"""
-    Multiplier(ops=(*,*))
+"""
+    $TYPEDEF
 
 Construts an `Multiplier` with input bus `input` and signs `signs`. `signs` is a tuplle of `*` and/or `/`. The output function `g` of `Multiplier` is of the form,
 ```math 
-    y = g(u, t) =  \prod_{j = 1}^n s_k u_k
+    y = g(u, t) =  \\prod_{j = 1}^n s_k u_k
 ```
 where `n` is the length of the `input`, ``s_k`` is the `k`th element of `signs`, ``u_k`` is the `k`th value of `input` and ``y`` is the value of the `output`. The default value of `signs` is all `*`.
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example 
 ```jldoctest
@@ -129,9 +148,13 @@ true
 ```
 """
 @def_static_system struct Multiplier{S, IP, OP, RO} <: AbstractStaticSystem
+    "Operators"
     ops::S = (*,*)
+    "Input port"
     input::IP = Inport(length(ops))
+    "Output port"
     output::OP = Outport()
+    "Readout function"
     readout::RO = (u, t, ops=ops) -> begin 
         ops = ops
         val = 1
@@ -143,14 +166,18 @@ true
 end
 
 
-@doc raw"""
-    Gain(input; gain=1.)
+"""
+    $TYPEDEF
 
 Constructs a `Gain` whose output function `g` is of the form 
 ```math 
     y = g(u, t) =  K u
 ```
 where ``K`` is `gain`, ``u`` is the value of `input` and `y` is the value of `output`.
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example 
 ```jldoctest
@@ -163,32 +190,47 @@ true
 ```
 """
[…]
