### 2021-05-09 984865f
Revised docstring of DAE systems #78
