### 2021-05-09 9d70a33
Revised docstrings in discrete systems. #78
