+++ src/components/systems/dynamicalsystems/discretesystems.jl
@@ -21,7 +21,8 @@ where `ex` is the expression to define to define a new AbstractDiscreteSystem co
     output::OP = output_default                 # mandatory field 
 end
 ```
-Here, `MyDiscreteSystem` has `N` parameters. `MyDiscreteSystem` is represented by the `righthandside` and `readout` function. `state`, `input` and `output` is the state, input port and output port of `MyDiscreteSystem`.
+Here, `MyDiscreteSystem` has `N` parameters. `MyDiscreteSystem` is represented by the `righthandside` and `readout` function.
+`state`, `input` and `output` is the state, input port and output port of `MyDiscreteSystem`.
 
 !!! warning 
     `righthandside` must have the signature 
@@ -77,9 +78,13 @@ end
 ##### Define Discrete system library
 
 """
-    DiscreteSystem(; righthandside, readout, state, input, output)
+    $TYPEDEF
 
-Constructs a generic discrete system 
+A generic discrete system 
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example 
 ```jldoctest
@@ -93,171 +98,247 @@ DiscreteSystem(righthandside:sfuncdiscrete, readout:ofuncdiscrete, state:[1.0], 
 ```
 """
 @def_discrete_system mutable struct DiscreteSystem{RH, RO, ST, IP, OP} <: AbstractDiscreteSystem
+    "Right-hand-side function"
     righthandside::RH
+    "Readout function"
     readout::RO 
+    "State"
     state::ST 
+    "Input. Expected to be an `Inport` or `Nothing`"
     input::IP 
+    "Output port"
     output::OP
 end
 
 
-@doc raw"""
-    DiscreteLinearSystem(input, output, modelargs=(), solverargs=(); 
-        A=fill(-1, 1, 1), B=fill(0, 1, 1), C=fill(1, 1, 1), D=fill(0, 1, 1), state=rand(size(A,1)), t=0., 
-        alg=ODEAlg, modelkwargs=NamedTuple(), solverkwargs=NamedTuple())
+"""
+    $TYPEDEF
 
-Constructs a `DiscreteLinearSystem` with `input` and `output`. `state` is the initial state and `t` is the time. `modelargs` and `modelkwargs` are passed into `ODEProblem` and `solverargs` and `solverkwargs` are passed into `solve` method of `DifferentialEquations`. `alg` is the algorithm to solve the differential equation of the system.
+Constructs a `DiscreteLinearSystem` with `input` and `output`. `state` is the initial state and `t` is the time. `modelargs`
+and `modelkwargs` are passed into `ODEProblem` and `solverargs` and `solverkwargs` are passed into `solve` method of
+`DifferentialEquations`. `alg` is the algorithm to solve the differential equation of the system.
 
 The `DiscreteLinearSystem` is represented by the following state and output equations.
 ```math
-\begin{array}{l}
-    \dot{x} = A x + B u \\[0.25cm]
+\\begin{array}{l}
+    \\dot{x} = A x + B u \\\\[0.25cm]
     y = C x + D u 
-\end{array}
+\\end{array}
 ```
[…]
