### 2021-05-09 fbce2a3
Revised docstrings of ODE systems #78
