+++ src/components/systems/dynamicalsystems/odesystems.jl
@@ -22,7 +22,8 @@ where `ex` is the expression to define to define a new AbstractODESystem compone
     output::OP = output_default                     # mandatory field 
 end
 ```
-Here, `MyODESystem` has `N` parameters. `MyODESystem` is represented by the `righthandside` and `readout` function. `state`, `input` and `output` is the state, input port and output port of `MyODESystem`.
+Here, `MyODESystem` has `N` parameters. `MyODESystem` is represented by the `righthandside` and `readout` function. `state`,
+`input` and `output` is the state, input port and output port of `MyODESystem`.
 
 !!! warning 
     `righthandside` must have the signature 
@@ -82,9 +83,13 @@ end
 #### Define ODE system library.
 
 """
-    ODESystem(;righthandside, readout, state, input, output) 
+    $TYPEDEF
 
 Constructs a generic ODE system.
+
+# Fields 
+
+    $TYPEDFIELDS
 
 # Example
 ```jldoctest
@@ -96,102 +101,144 @@ julia> ds.state
 ```
 """
 @def_ode_system mutable struct ODESystem{RH, RO, ST, IP, OP} <: AbstractODESystem 
+    "Right-hand-side function"
     righthandside::RH 
+    "Readout function"
     readout::RO 
+    "State"
     state::ST 
+    "Input. Expected to an `Inport` or `Nothing`"
     input::IP 
+    "Output port"
     output::OP
 end
 
-@doc raw"""
-    ContinuousLinearSystem(input, output, modelargs=(), solverargs=(); 
-        A=fill(-1, 1, 1), B=fill(0, 1, 1), C=fill(1, 1, 1), D=fill(0, 1, 1), state=rand(size(A,1)), t=0., 
-        alg=ODEAlg, modelkwargs=NamedTuple(), solverkwargs=NamedTuple())
+"""
+    $TYPEDEF
 
-Constructs a `ContinuousLinearSystem` with `input` and `output`. `state` is the initial state and `t` is the time. `modelargs` and `modelkwargs` are passed into `ODEProblem` and `solverargs` and `solverkwargs` are passed into `solve` method of `DifferentialEquations`. `alg` is the algorithm to solve the differential equation of the system.
+Constructs a `ContinuousLinearSystem` with `input` and `output`. `state` is the initial state and `t` is the time.
+`modelargs` and `modelkwargs` are passed into `ODEProblem` and `solverargs` and `solverkwargs` are passed into `solve` method
+of `DifferentialEquations`. `alg` is the algorithm to solve the differential equation of the system.
 
 The `ContinuousLinearSystem` is represented by the following state and output equations.
 ```math
-\begin{array}{l}
-    \dot{x} = A x + B u \\[0.25cm]
+\\begin{array}{l}
+    \\dot{x} = A x + B u \\\\[0.25cm]
     y = C x + D u 
-\end{array}
+\\end{array}
 ```
 where ``x`` is `state`. `solver` is used to solve the above differential equation.
+
+# Fields    
+
+    $TYPEDFIELDS
 """
 @def_ode_system mutable struct ContinuousLinearSystem{IP, OP, RH, RO} <: AbstractODESystem
[…]
