### 2021-05-09 390d42c
Revised docstring of RODE and SDE systems Fixes #78
