### 2021-05-09 30ac833
Relaxed field types in source.jl #81
