### 2021-05-09 2702b98
Relaxed types of the fields of  types in odesystems.jl. #81
