+++ src/components/systems/dynamicalsystems/daesystems.jl
@@ -25,7 +25,9 @@ where `ex` is the expression to define to define a new AbstractDAESystem compone
     output::OP = output_default                 # mandatory field
 end
 ```
-Here, `MyDAESystem` has `N` parameters. `MyDAESystem` is represented by the `righthandside` and `readout` function. `state`, 'stateder`, `diffvars`, `input` and `output` is the initial state, initial value of differential variables, vector signifing differetial variables, input port and output port of `MyDAESystem`.
+Here, `MyDAESystem` has `N` parameters. `MyDAESystem` is represented by the `righthandside` and `readout` function. `state`,
+'stateder`, `diffvars`, `input` and `output` is the initial state, initial value of differential variables, vector signifing
+differetial variables, input port and output port of `MyDAESystem`.
 
 !!! warning 
     `righthandside` must have the signature 
@@ -104,7 +106,13 @@ julia> DAESystem(righthandside=sfuncdae, readout=ofuncdae, state=x0, input=nothi
 DAESystem(righthandside:sfuncdae, readout:ofuncdae, state:[1.0, -1.0], t:0.0, input:nothing, output:Outport(numpins:1, eltype:Outpin{Float64}))
 ```
 """
-@def_dae_system mutable struct DAESystem{RH, RO, ST, IP, OP} <: AbstractDAESystem
+@def_dae_system mutable struct DAESystem{RH, 
+                                         RO, 
+                                         ST <: AbstractVector{<:Real}, 
+                                         SD <: AbstractVector{<:Real}, 
+                                         DV <: AbstractVector{<:Bool}, 
+                                         IP <: Union{<:Inport, <:Nothing}, 
+                                         OP <: Union{<:Outport,<:Nothing}} <: AbstractDAESystem
     "Right-hand-side function."
     righthandside::RH 
     "Readout function"
@@ -112,9 +120,9 @@ DAESystem(righthandside:sfuncdae, readout:ofuncdae, state:[1.0, -1.0], t:0.0, in
     "Initial state"
     state::ST 
     "Initial state derivative"
-    stateder::ST 
+    stateder::SD 
     "The `true` elements of `diffvars` correspond to differential variables"
-    diffvars::Vector{Bool}
+    diffvars::DV
     "Input. Expected to be an `Inport` of `Nothing`"
     input::IP 
     "Output port"
@@ -134,13 +142,22 @@ Constructs a Robertson systme with the dynamcis
 \end{array}
 ```
 """
-@def_dae_system mutable struct RobertsonSystem{RH, RO, IP, OP} <: AbstractDAESystem 
+@def_dae_system mutable struct RobertsonSystem{T1 <: Real,
+                                               T2 <: Real, 
+                                               T3 <: Real, 
+                                               RH, 
+                                               RO, 
+                                               IP <: Union{<:Inport, <:Nothing}, 
+                                               OP <: Union{<:Outport,<:Nothing},
+                                               T4 <: AbstractVector{<:Real},
+                                               T5 <: AbstractVector{<:Real},
+                                               T6 <: AbstractVector{<:Bool}} <: AbstractDAESystem 
     "k1"
-    k1::Float64 = 0.04   
+    k1::T1 = 0.04   
     "k2"
-    k2::Float64 = 3e7 
+    k2::T2 = 3e7 
     "k3"
-    k3::Float64 = 1e4 
+    k3::T3 = 1e4 
     "Right-hand-side function"
     righthandside::RH = function robertsonrhs(out, dx, x, u, t)
[…]
