+++ src/components/systems/dynamicalsystems/discretesystems.jl
@@ -97,7 +97,11 @@ DiscreteSystem(righthandside:sfuncdiscrete, readout:ofuncdiscrete, state:[1.0], 
 
 ```
 """
-@def_discrete_system mutable struct DiscreteSystem{RH, RO, ST, IP, OP} <: AbstractDiscreteSystem
+@def_discrete_system mutable struct DiscreteSystem{RH, 
+                                                   RO, 
+                                                   ST <: AbstractVector{<:Real}, 
+                                                   IP <: Union{<:Inport, <:Nothing},
+                                                   OP <: Union{<:Outport,<:Nothing}} <: AbstractDiscreteSystem
     "Right-hand-side function"
     righthandside::RH
     "Readout function"
@@ -131,21 +135,29 @@ where ``x`` is `state`. `solver` is used to solve the above differential equatio
 
     $TYPEDFIELDS
 """
-@def_discrete_system mutable struct DiscreteLinearSystem{IP, OP, RH, RO} <: AbstractDiscreteSystem
+@def_discrete_system mutable struct DiscreteLinearSystem{T1 <: AbstractMatrix{<:Real}, 
+                                                         T2 <: AbstractMatrix{<:Real}, 
+                                                         T3 <: AbstractMatrix{<:Real}, 
+                                                         T4 <: AbstractMatrix{<:Real}, 
+                                                         IP <: Union{<:Inport, <:Nothing}, 
+                                                         OP <: Union{<:Outport,<:Nothing}, 
+                                                         ST <: AbstractVector{<:Real}, 
+                                                         RH, 
+                                                         RO} <: AbstractDiscreteSystem
     "A"
-    A::Matrix{Float64} = fill(-1., 1, 1)
+    A::T1 = fill(-1., 1, 1)
     "B"
-    B::Matrix{Float64} = fill(0., 1, 1)
+    B::T2 = fill(0., 1, 1)
     "C"
-    C::Matrix{Float64} = fill(1., 1, 1)
+    C::T3 = fill(1., 1, 1)
     "D"
-    D::Matrix{Float64} = fill(-1., 1, 1)
+    D::T4 = fill(-1., 1, 1)
     "Input. Expected to an `Inport` or `Nothing`"
     input::IP = Inport(1)
     "Output port"
     output::OP = nothing
     "State"
-    state::Vector{Float64} = rand(size(A, 1))
+    state::ST = rand(size(A, 1))
     "Right-hand-side function"
     righthandside::RH = input === nothing ? (dx, x, u, t) -> (dx .= A * x) : 
         (dx, x, u, t) -> (dx .= A * x + B * map(ui -> ui(t), u.itp))
@@ -170,13 +182,20 @@ Constructs a `Henon` system evolving with the dynamics
 
     $TYPEDFIELDS
 """
-@def_discrete_system mutable struct HenonSystem{RH, RO, IP, OP} <: AbstractDiscreteSystem
+@def_discrete_system mutable struct HenonSystem{T1 <: Real,
+                                                T2 <: Real,  
+                                                T3 <: Real,  
+                                                RH, 
+                                                RO, 
+                                                ST <: AbstractVector{<:Real},
+                                                IP <: Union{<:Inport, <:Nothing}, 
+                                                OP <: Union{<:Outport,<:Nothing}} <: AbstractDiscreteSystem
     "α"
-    α::Float64 = 1.4 
+    α::T1 = 1.4 
     "β"
-    β::Float64 = 0.3 
+    β::T2 = 0.3 
     "γ"
-    γ::Float64 = 1.
+    γ::T3 = 1.
[…]

[…]
