### 2021-05-10 ed726ec
Added methods to resize sink buffers.
