### 2021-05-10 5214534
Returned mdoel from resize_sink_buffers! function.
