### 2021-06-08 #71 @YingboMa
Sorry for the late reply. I only have a few minor comments on this paper, and overall, I think this paper is really solid. Though, I am not familiar with causal modeling either.

1. Could you write a bit more on the connection between the sampling by the model clock and the time stepping of the differential equation solvers?
2. There are a few times where the space before the open parenthesis is missing:
![image](https://user-images.githubusercontent.com/17304743/121257734-09bfd980-c87c-11eb-941a-097465a5cae9.png)
3. Do the strategies of breaking up algebraic loops suggested in Section 2.4 compromise the accuracy of the model?
