+++ paper/paper.tex
@@ -27,7 +27,7 @@ Numerical simulations can be expressed as solving the mathematical equations der
 
 Many simulation environments are available for numerical analysis of systems\cite{elmqvist1978structured,nytsch2006advanced,zimmer2008introducing,mosterman2002hybrsim,van2001variables,giorgidze2009higher,pfeiffer2012pysimulator,simulink}. They are capable of allowing simulations that are represented by ordinary differential equations and differential algebraic equations, mostly. This is restrictive given the variety of mathematical equations that can be derived from the modeling\cite{rackauckas2017differentialequations}. Besides, many of the existing simulation environments lack modern computational methods such as parallel computing.
 
-In this study, Causal.jl, a modeling and simulation framework for causal models, is introduced \cite{causal}. The aim is to model large scale complex system networks easily and to provide fast and effective simulations. For this purpose, Julia, an open-source, high level, general purpose dynamical programming language designed for high-performance numerical analysis and computational science, has been used. Although Julia is a dynamical language, owing to its Just-in-Time(JIT) compiler developed on Low Level Virtual Machine(LLVM), it can reach the high-speed performance of static languages such as C\cite{bezanson2017julia,julialang}. It supports various parallel computing techniques at thread and process levels. In addition to Julia's standard library, numerous specialized packages developed for different fields such as data science, scientific computing, are also available. Julia's high-speed performance and parallel computing support are essential in meeting the need to design a  fast and effective simulation environment. Julia's syntax can be enlarged purposefully using its metaprogramming support. The analyzes scope of the simulation framework can be extended with new plugins that can be easily defined. It is possible to analyze discrete or continuous-time, static, or dynamical systems. In particular, it is possible to simulate dynamical systems modeled by ordinary, random ordinary, stochastic, delay differential, differential-algebraic, and/or discrete difference equations simultaneously. Unlike its counterparts, the models do not evolve at once for the whole simulation duration. Instead, the model components evolve between sampling time intervals individually. While individual evolution of the components enables the simulation of systems consisting of components represented by different mathematical models, parallel evolution of components increases the simulation performance.
+In this study, Causal.jl, a modeling and simulation framework for causal models, is introduced \cite{causal}. The aim is to model large scale complex system networks easily and to provide fast and effective simulations. For this purpose, Julia, an open-source, high level, general purpose dynamical programming language designed for high-performance numerical analysis and computational science, has been used. Although Julia is a dynamical language, owing to its Just-in-Time (JIT) compiler developed on Low Level Virtual Machine (LLVM), it can reach the high-speed performance of static languages such as C\cite{bezanson2017julia,julialang}. It supports various parallel computing techniques at thread and process levels. In addition to Julia's standard library, numerous specialized packages developed for different fields such as data science, scientific computing, are also available. Julia's high-speed performance and parallel computing support are essential in meeting the need to design a  fast and effective simulation environment. Julia's syntax can be enlarged purposefully using its metaprogramming support. The analyzes scope of the simulation framework can be extended with new plugins that can be easily defined. It is possible to analyze discrete or continuous-time, static, or dynamical systems. In particular, it is possible to simulate dynamical systems modeled by ordinary, random ordinary, stochastic, delay differential, differential-algebraic, and/or discrete difference equations simultaneously. Unlike its counterparts, the models do not evolve at once for the whole simulation duration. Instead, the model components evolve between sampling time intervals individually. While individual evolution of the components enables the simulation of systems consisting of components represented by different mathematical models, parallel evolution of components increases the simulation performance.
 
 
 
[…]
