### 2021-07-04 #71 @zekeriyasari
Hi @matbesancon

I am done with the updates. I have just committed the updates to the repository. Similar to the previous case, in the commit is also included a jupyter notebook titled [answers2.ipynb](https://github.com/zekeriyasari/Causal.jl/blob/juliacon/paper/answers2.ipynb) that includes the specific answers of questions/comments of the review by @YingboMa
