### 2021-07-04 #71 @zekeriyasari
@matbesancon I just added two files in `paper` folder: `answers2.ipynb` and `algebraicloop.svg` (this is included in `answers2.ipynb`). Nothing else. And, I can compile `paper.tex` using ` latexmk -bibtex -pdf paper.tex` without a problem.
