### 2021-07-05 #71 @zekeriyasari
Hi, @matbesancon

I deleted all the files related to the answers to the reviewers from `paper` folder. But, still, pdf seems not to be generated. I could not understand what goes wrong. Anything else I can do on my side?
