### 2021-07-06 #71 @arfon
> no I'm not sure. @arfon do you have an idea what could cause this?

It looks like @whedon was removed from this repo and https://github.com/JuliaCon/proceedings-papers a couple of weeks ago, perhaps when a 2FA requirement was added to the JuliaCon GitHub organization? This means Whedon can no longer push papers to the https://github.com/JuliaCon/proceedings-papers repo.
