+++ paper/paper.bib
@@ -1,149 +1,156 @@
 @article{elmqvist1978structured,
-  title={A structured model language for large continuous systems},
-  author={Elmqvist, Hilding},
-  journal={PhD Thesis TFRT-1015},
-  year={1978},
-  publisher={Department of Automatic Control, Lund Institute of Technology (LTH)}
+  title     = {A structured model language for large continuous systems},
+  author    = {Elmqvist, Hilding},
+  journal   = {PhD Thesis TFRT-1015},
+  year      = {1978},
+  publisher = {Department of Automatic Control, Lund Institute of Technology (LTH)}
 }
 @inproceedings{nytsch2006advanced,
-  title={Advanced modeling and simulation techniques in MOSILAB: A system development case study},
-  author={Nytsch-Geusen, Christoph and Ernst, Thilo and Nordwig, Andr{\'e} and Schwarz, Peter and Schneider, Peter and Vetter, Matthias and Wittwer, Christof and Holm, Andreas and Nouidui, Thierry and Leopold, J{\"u}rgen and others},
-  booktitle={Proceedings of the 5th International Modelica Conference, Arsenal Research, Wien},
-  year={2006}
+  title     = {Advanced modeling and simulation techniques in MOSILAB: A system development case study},
+  author    = {Nytsch-Geusen, Christoph and Ernst, Thilo and Nordwig, Andr{\'e} and Schwarz, Peter and Schneider, Peter and Vetter, Matthias and Wittwer, Christof and Holm, Andreas and Nouidui, Thierry and Leopold, J{\"u}rgen and others},
+  booktitle = {Proceedings of the 5th International Modelica Conference, Arsenal Research, Wien},
+  year      = {2006}
 }
 @inproceedings{zimmer2008introducing,
-  title={Introducing Sol: A general methodology for equation-based modeling of variable-structure systems},
-  author={Zimmer, Dirk},
-  booktitle={Proceedings of the 6th International Modelica Conference},
-  pages={47--56},
-  year={2008},
-  organization={Citeseer}
+  title        = {Introducing Sol: A general methodology for equation-based modeling of variable-structure systems},
+  author       = {Zimmer, Dirk},
+  booktitle    = {Proceedings of the 6th International Modelica Conference},
+  pages        = {47--56},
+  year         = {2008},
+  organization = {Citeseer}
 }
 @article{mosterman2002hybrsim,
-  title={HYBRSIM—a modelling and simulation environment for hybrid bond graphs},
-  author={Mosterman, Pieter J},
-  journal={Proceedings of the Institution of Mechanical Engineers, Part I: Journal of Systems and Control Engineering},
-  volume={216},
-  number={1},
-  pages={35--46},
-  year={2002},
-  publisher={SAGE Publications Sage UK: London, England}
+  title     = {HYBRSIM—a modelling and simulation environment for hybrid bond graphs},
+  author    = {Mosterman, Pieter J},
+  journal   = {Proceedings of the Institution of Mechanical Engineers, Part I: Journal of Systems and Control Engineering},
+  volume    = {216},
+  number    = {1},
+  pages     = {35--46},
+  year      = {2002},
+  doi       = {10.1243/0959651021541417},
+  publisher = {SAGE Publications Sage UK: London, England}
 }
 @inproceedings{van2001variables,
-  title={Variables and equations in hybrid systems with structural changes},
-  author={Van Beek, DA},
-  booktitle={Proc. 13th European Simulation Symposium, Marseille},
-  pages={30--34},
-  year={2001},
-  organization={Citeseer}
[…]
