+++ paper/paper.tex
@@ -83,7 +83,7 @@ The data transferred to a port are transferred to its connections. The data tran
 
 In the modeling approach adopted, the components reading data from a connection are driven by other components writing data to the connection. Therefore, all of the model's connections must be readable and writable so that data can flow through the connections. This necessitates that all the connections of the model must be connected to a component from both ends. Otherwise, the simulation gets stuck and does not terminate.
 
-There exist other tools developed in Julia for parallel execution of computations. Dagger.jl is such an example of those tools \cite{dagger}. Dagger.jl distributes the computations that can be represented by a directed acyclic graph (DAG) to available worker processes to execute them in parallel. A DAG is a directed graph with no directed cycles (loops). The DAG modeling approach brings restrictions to the simulation of the models that include loops, which is a case commonly encountered e.g. in feedback control systems. However, Causal.jl is capable of simulating the systems that include loops in their models. 
+There exist other tools developed in Julia for parallel execution of computations. Dagger.jl is such an example of those tools \cite{dagger}. Dagger.jl distributes the computations that can be represented by a directed acyclic graph (DAG) to available worker processes to execute them in parallel. A DAG is a directed graph with no directed cycles (loops). The DAG modeling approach brings restrictions to the simulation of the models that include loops, which is a case commonly encountered e.g. in feedback control systems. However, Causal.jl is capable of simulating the systems that include loops in their models.
 
 
 \subsection{Simulation}
@@ -150,21 +150,21 @@ In Causal.jl, it is possible to simulate models that include components represen
     \begin{split}
         dx_2 &= (a (y_1 - x_1) + u_1) dt + \eta dW_t \\
         dy_2 &= ((c - 1) x_1 + c y_1 - x_1 z_1)dt + \eta dW_t \\
-        dz_2 &= (x_1 y_1 - b z_1 + u_3) dt + \eta dW_t 
+        dz_2 &= (x_1 y_1 - b z_1 + u_3) dt + \eta dW_t
     \end{split}
     \label{eq: coupled sde}
 \end{equation}
-\texttt{gain} is a static system whose input output relation is given in (\ref{eq: gain}) where $u, y, \epsilon$ are the input, output and gain coefficient of the system. 
+\texttt{gain} is a static system whose input output relation is given in (\ref{eq: gain}) where $u, y, \epsilon$ are the input, output and gain coefficient of the system.
 \begin{equation}
     y = \epsilon u
     \label{eq: gain}
 \end{equation}
-\texttt{writer1} and \texttt{writer2} are used to record the outputs of \texttt{odeds} and \texttt{sdeds}, respectively, while \texttt{writer3} is used to record the fast Fourier transform (FFT) of the data flowing out of \texttt{odeds}\cite{proakis2004digital}. 
+\texttt{writer1} and \texttt{writer2} are used to record the outputs of \texttt{odeds} and \texttt{sdeds}, respectively, while \texttt{writer3} is used to record the fast Fourier transform (FFT) of the data flowing out of \texttt{odeds}\cite{proakis2001digital}.
 
[…]
