+++ paper/ref.bib
@@ -27,6 +27,7 @@
   number    = {1},
   pages     = {35--46},
   year      = {2002},
+  doi       = {10.1243/0959651021541417},
   publisher = {SAGE Publications Sage UK: London, England}
 }
 @inproceedings{van2001variables,
@@ -44,6 +45,7 @@
   number       = {043},
   pages        = {208--218},
   year         = {2009},
+  doi          = {10.3384/ecp09430137},
   organization = {Link{\"o}ping University Electronic Press}
 }
 @article{pfeiffer2012pysimulator,
@@ -68,6 +70,7 @@
   number    = {1},
   pages     = {65--98},
   year      = {2017},
+  doi       = {10.1137/141000671},
   publisher = {SIAM}
 }
 @book{matei2012modeling,
@@ -84,6 +87,7 @@
   number    = {1},
   pages     = {33--42},
   year      = {2001},
+  doi       = {10.1109/72.896794},
   publisher = {IEEE}
 }
 @book{van2000introduction,
@@ -103,11 +107,11 @@
   year      = {1998},
   publisher = {Nature Publishing Group}
 }
-@article{proakis2004digital,
-  title   = {Digital signal processing},
-  author  = {Proakis, John G and Manolakis, Dimitris G},
-  journal = {PHI Publication: New Delhi, India},
-  year    = {2004}
+@book{proakis2001digital,
+  title     = {Digital signal processing: principles algorithms and applications},
+  author    = {Proakis, John G},
+  year      = {2001},
+  publisher = {Pearson Education India}
 }
 @misc{simulink,
   title        = {Simulink-Simulation and Model-Based Design},
