### 2022-01-27 #71 @matbesancon
Hi @zekeriyasari, there is an issue with the formatting, can you remove the backslashes from the title in the yml file and let us know.
The problem is that this is also copied to the paper title: https://proceedings.juliacon.org/papers/10.21105/jcon.00071
and to the bibtex entry
Please let us know when this is done
